# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Notes App API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Folder {
  id: number;
  name: string;
  /** @nullable */
  parentId: number | null;
  /** @nullable */
  color: string | null;
  /** @nullable */
  icon: string | null;
  sortOrder: number;
  createdAt: string;
  updatedAt: string;
}

export interface CreateFolderBody {
  name: string;
  /** @nullable */
  parentId?: number | null;
  /** @nullable */
  color?: string | null;
  /** @nullable */
  icon?: string | null;
  sortOrder?: number;
}

export interface UpdateFolderBody {
  name?: string;
  /** @nullable */
  parentId?: number | null;
  /** @nullable */
  color?: string | null;
  /** @nullable */
  icon?: string | null;
  sortOrder?: number;
}

export interface Note {
  id: number;
  title: string;
  content: string;
  /** @nullable */
  contentText: string | null;
  /** @nullable */
  folderId: number | null;
  tags: string[];
  pinned: boolean;
  favorite: boolean;
  /** @nullable */
  coverImage: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreateNoteBody {
  title: string;
  content: string;
  /** @nullable */
  contentText?: string | null;
  /** @nullable */
  folderId?: number | null;
  tags?: string[];
  pinned?: boolean;
  favorite?: boolean;
  /** @nullable */
  coverImage?: string | null;
}

export interface UpdateNoteBody {
  title?: string;
  content?: string;
  /** @nullable */
  contentText?: string | null;
  /** @nullable */
  folderId?: number | null;
  tags?: string[];
  pinned?: boolean;
  favorite?: boolean;
  /** @nullable */
  coverImage?: string | null;
}

export interface MoveNoteBody {
  /** @nullable */
  folderId: number | null;
}

export type AiCompleteBodyProvider =
  (typeof AiCompleteBodyProvider)[keyof typeof AiCompleteBodyProvider];

export const AiCompleteBodyProvider = {
  openai: "openai",
  anthropic: "anthropic",
  google: "google",
} as const;

export interface AiCompleteBody {
  provider: AiCompleteBodyProvider;
  apiKey: string;
  model: string;
  prompt: string;
  /** @nullable */
  systemPrompt?: string | null;
  /** @nullable */
  noteContext?: string | null;
}

export interface AiCompleteResponse {
  result: string;
  /** @nullable */
  tokensUsed: number | null;
}

export type GetNotesParams = {
  /**
   * @nullable
   */
  folderId?: number | null;
  /**
   * @nullable
   */
  search?: string | null;
  /**
   * @nullable
   */
  pinned?: boolean | null;
  /**
   * @nullable
   */
  favorite?: boolean | null;
  /**
   * @nullable
   */
  tag?: string | null;
  /**
   * @nullable
   */
  sortBy?: GetNotesSortBy;
  /**
   * @nullable
   */
  sortDir?: GetNotesSortDir;
};

export type GetNotesSortBy =
  | (typeof GetNotesSortBy)[keyof typeof GetNotesSortBy]
  | null;

export const GetNotesSortBy = {
  updatedAt: "updatedAt",
  createdAt: "createdAt",
  title: "title",
} as const;

export type GetNotesSortDir =
  | (typeof GetNotesSortDir)[keyof typeof GetNotesSortDir]
  | null;

export const GetNotesSortDir = {
  asc: "asc",
  desc: "desc",
} as const;