# lib/api-zod/src/generated/types/aiCompleteBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Notes App API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AiCompleteBodyProvider } from "./aiCompleteBodyProvider";

export interface AiCompleteBody {
  provider: AiCompleteBodyProvider;
  apiKey: string;
  model: string;
  prompt: string;
  /** @nullable */
  systemPrompt?: string | null;
  /** @nullable */
  noteContext?: string | null;
}