# lib/api-zod/src/generated/types/aiCompleteResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Notes App API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AiCompleteResponse {
  result: string;
  /** @nullable */
  tokensUsed: number | null;
}