# lib/api-zod/src/generated/types/note.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Notes App API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Note {
  id: number;
  title: string;
  content: string;
  /** @nullable */
  contentText: string | null;
  /** @nullable */
  folderId: number | null;
  tags: string[];
  pinned: boolean;
  favorite: boolean;
  /** @nullable */
  coverImage: string | null;
  createdAt: Date;
  updatedAt: Date;
}