# lib/api-zod/src/generated/types/updateFolderBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Notes App API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateFolderBody {
  name?: string;
  /** @nullable */
  parentId?: number | null;
  /** @nullable */
  color?: string | null;
  /** @nullable */
  icon?: string | null;
  sortOrder?: number;
}